requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
gradio==5.23.2
transformers==4.46.3
torch>=2.4.0
accelerate==1.1.1
peft==0.13.2
datasets==3.1.0
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
python-dotenv==1.0.1
